+++ plugins/Feedback/tests/Integration/ControllerTest.php
@@ -36,8 +36,6 @@ class ControllerTest extends IntegrationTestCase
             'user1',
             'a98732d98732',
             'user1@example.com',
-            'user1',
-            'ab9879dc23876f19',
             '2019-03-03'
         );
         FakeAccess::$identity = 'user1';