+++ plugins/Insights/tests/Unit/FilterOrderByTest.php
@@ -6,7 +6,7 @@
  * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
  */
 
-namespace Piwik\Plugins\Insights\tests;
+namespace Piwik\Plugins\Insights\tests\Unit;
 
 use Piwik\DataTable;
 use Piwik\DataTable\Row;