+++ tests/PHPUnit/Fixtures/JSTrackingUIFixture.php
@@ -9,6 +9,8 @@
 
 namespace Piwik\Tests\Fixtures;
 
+use Piwik\Container\StaticContainer;
+use Piwik\Plugins\CustomJsTracker\TrackerUpdater;
 use Piwik\Plugins\GeoIp2\LocationProvider\GeoIp2\Php;
 use Piwik\Plugins\PrivacyManager\IPAnonymizer;
 use Piwik\Plugins\UserCountry\LocationProvider;
@@ -25,6 +27,9 @@ class JSTrackingUIFixture extends Fixture
         self::installAndActivatePlugins($this->getTestEnvironment());
         self::updateDatabase();
 
+        $trackerUpdater = StaticContainer::get('Piwik\Plugins\CustomJsTracker\TrackerUpdater');
+        $trackerUpdater->update();
+
         // for proper geolocation
         LocationProvider::setCurrentProvider(Php::ID);
         IPAnonymizer::deactivate();
@@ -38,6 +43,7 @@ class JSTrackingUIFixture extends Fixture
             'loadRealTranslations' => 1,
         );
         $this->extraPluginsToLoad = array(
+            'CustomJsTracker',
             'ExampleTracker',
         );